Relocate Metadata record classes inside fccloudapi
